+++ code/Player.h
@@ -8,7 +8,7 @@ class Player : public GameObject
 public:
   Player();
   // x location, y location, pointer to display
-  Player(int xl, int yl, Display* display);
+  Player(double xl, double yl, Display* display);
   ~Player();
 
   void handleEvent(SDL_Event* e);
